integration tests: refactor common into utils mod
As retries.rs contained some features that can certainly be useful in
other tests, they were extracted into a new, shared mod "utils".
