+++ scylla/tests/retries.rs
@@ -1,103 +1,18 @@
-use futures::Future;
-use itertools::Itertools;
-use scylla::load_balancing::LoadBalancingPolicy;
+mod utils;
+
 use scylla::query::Query;
 use scylla::retry_policy::FallthroughRetryPolicy;
 use scylla::speculative_execution::SimpleSpeculativeExecutionPolicy;
 use scylla::transport::session::Session;
 use scylla::SessionBuilder;
-use std::collections::HashMap;
-use std::env;
-use std::net::SocketAddr;
-use std::str::FromStr;
 use std::sync::Arc;
 use std::time::Duration;
-use tracing::{info, instrument::WithSubscriber};
+use tracing::info;
+use utils::test_with_3_node_cluster;
 
 use scylla_proxy::{
-    Condition, Node, Proxy, ProxyError, Reaction, RequestOpcode, RequestReaction, RequestRule,
-    RunningProxy, ShardAwareness, WorkerError,
+    Condition, ProxyError, Reaction, RequestOpcode, RequestReaction, RequestRule, WorkerError,
 };
-
-fn init_logger() {
-    let _ = tracing_subscriber::fmt::fmt()
-        .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
-        .without_time()
-        .try_init();
-}
-
-#[derive(Debug)]
-struct FixedOrderLoadBalancer;
-impl LoadBalancingPolicy for FixedOrderLoadBalancer {
-    fn plan<'a>(
-        &self,
-        _statement: &scylla::load_balancing::Statement,
-        cluster: &'a scylla::transport::ClusterData,
-    ) -> scylla::load_balancing::Plan<'a> {
-        Box::new(
-            cluster
-                .get_nodes_info()
-                .clone()
-                .into_iter()
-                .sorted_by(|node1, node2| Ord::cmp(&node1.address, &node2.address)),
-        )
-    }
-
-    fn name(&self) -> String {
-        "FixedOrderLoadBalancer".to_string()
-    }
-}
-
-async fn test_with_3_node_cluster<F, Fut>(
-    first_proxy_node_addr_last_octet: u16,
-    test: F,
-) -> Result<(), ProxyError>
-where
-    F: FnOnce([String; 3], HashMap<SocketAddr, SocketAddr>, RunningProxy) -> Fut,
-    Fut: Future<Output = RunningProxy>,
-{
-    init_logger();
-    let real1_uri = env::var("SCYLLA_URI").unwrap_or_else(|_| "127.0.0.1:9042".to_string());
-    let proxy1_uri = format!("127.0.0.{}:9042", first_proxy_node_addr_last_octet);
-    let real2_uri = env::var("SCYLLA_URI2").unwrap_or_else(|_| "127.0.0.2:9042".to_string());
-    let proxy2_uri = format!("127.0.0.{}:9042", first_proxy_node_addr_last_octet + 1);
-    let real3_uri = env::var("SCYLLA_URI3").unwrap_or_else(|_| "127.0.0.3:9042".to_string());
-    let proxy3_uri = format!("127.0.0.{}:9042", first_proxy_node_addr_last_octet + 2);
-
-    let real1_addr = SocketAddr::from_str(real1_uri.as_str()).unwrap();
-    let proxy1_addr = SocketAddr::from_str(proxy1_uri.as_str()).unwrap();
-    let real2_addr = SocketAddr::from_str(real2_uri.as_str()).unwrap();
-    let proxy2_addr = SocketAddr::from_str(proxy2_uri.as_str()).unwrap();
-    let real3_addr = SocketAddr::from_str(real3_uri.as_str()).unwrap();
-    let proxy3_addr = SocketAddr::from_str(proxy3_uri.as_str()).unwrap();
-
-    let proxy = Proxy::new(
-        [
-            (proxy1_addr, real1_addr),
-            (proxy2_addr, real2_addr),
-            (proxy3_addr, real3_addr),
-        ]
-        .map(|(proxy_addr, real_addr)| {
-            Node::builder()
-                .real_address(real_addr)
-                .proxy_address(proxy_addr)
-                .shard_awareness(ShardAwareness::QueryNode)
-                .build()
-        }),
-    );
-
-    let translation_map = proxy.translation_map();
-    let running_proxy = proxy.run().with_current_subscriber().await.unwrap();
-
-    let running_proxy = test(
-        [proxy1_uri, proxy2_uri, proxy3_uri],
-        translation_map,
-        running_proxy,
-    )
-    .await;
-
-    running_proxy.finish().await
-}
 
 #[tokio::test]
 #[ntest::timeout(5000)]
